Reverting changes to uint64_t

+++ api.cpp
@@ -12,7 +12,6 @@
 #include <map>
 #include <set>
 #include <string>
-#include <cstdint> //has type 'uint64_t'
 
 using std::pair;
 using std::map;
@@ -20,9 +19,9 @@ using std::set;
 using std::string;
 
 
-uint64_t event_add_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
+int event_add_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
 	//Call graph function
-	uint64_t result = (*graph).add_node(node_id);
+	int result = (*graph).add_node(node_id);
 
 	//Send HTTP reply
 	if(result == 1){
@@ -49,9 +48,9 @@ uint64_t event_add_node(Graph *graph, struct mg_connection *nc, uint64_t node_id
 	return 1;
 }
 
-uint64_t event_add_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
+int event_add_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
 	//Call graph function
-	uint64_t result = (*graph).add_edge(node_a_id, node_b_id);
+	int result = (*graph).add_edge(node_a_id, node_b_id);
 
 	//Send HTTP reply
 	if(result == 1){
@@ -80,9 +79,9 @@ uint64_t event_add_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_
 	return 1;
 }
 
-uint64_t event_remove_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
+int event_remove_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
 	//Call graph function
-	uint64_t result = (*graph).remove_node(node_id);
+	int result = (*graph).remove_node(node_id);
 
 	//Send HTTP reply
 	if(result == 1){
@@ -107,9 +106,9 @@ uint64_t event_remove_node(Graph *graph, struct mg_connection *nc, uint64_t node
 	return 1;
 }
 
-uint64_t event_remove_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
+int event_remove_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
 	//Call graph function
-	uint64_t result = (*graph).remove_edge(node_a_id, node_b_id);
+	int result = (*graph).remove_edge(node_a_id, node_b_id);
 
 	//Send HTTP reply
 	if(result == 1){
@@ -134,9 +133,9 @@ uint64_t event_remove_edge(Graph *graph, struct mg_connection *nc, uint64_t node
 	return 1;
 }
 
-uint64_t event_get_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
+int event_get_node(Graph *graph, struct mg_connection *nc, uint64_t node_id){
 	//Call graph function
-	uint64_t result = (*graph).get_node(node_id);
+	int result = (*graph).get_node(node_id);
 
 	//Send HTTP reply
 	mg_send_head(nc, 200, -1, NULL);
@@ -158,9 +157,9 @@ uint64_t event_get_node(Graph *graph, struct mg_connection *nc, uint64_t node_id
 	return 1;
 }
 
-uint64_t event_get_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
+int event_get_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
 	//Call graph function
-	uint64_t result = (*graph).get_edge(node_a_id, node_b_id);
+	int result = (*graph).get_edge(node_a_id, node_b_id);
 
 	//Send HTTP reply
 	if(result != 2){
@@ -189,9 +188,9 @@ uint64_t event_get_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_
 	return 1;
 }
 
-uint64_t event_get_neighbors(Graph *graph, struct mg_connection *nc, uint64_t node_id){
+int event_get_neighbors(Graph *graph, struct mg_connection *nc, uint64_t node_id){
 	//Call graph function
-	pair<uint64_t, set<uint64_t> > result = (*graph).get_neighbors(node_id);
+	pair<int, set<uint64_t> > result = (*graph).get_neighbors(node_id);
 
 	//Send HTTP reply
 	if(result.first == 1){
@@ -235,9 +234,9 @@ uint64_t event_get_neighbors(Graph *graph, struct mg_connection *nc, uint64_t no
 	return 1;
 }
 
-uint64_t event_shortest_path(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
+int event_shortest_path(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id){
 	//Call graph function
-	pair<uint64_t, uint64_t> result = (*graph).shortest_path(node_a_id, node_b_id);
+	pair<int, uint64_t> result = (*graph).shortest_path(node_a_id, node_b_id);
 
 	//Send HTTP reply
 	if(result.first == 1){
